1. 돌연변이 확률을 float로 변경 2. 정수 돌연변이가 과하게 일어나지 않도록 개선 3. 정수 교체 과정에서 쉬프트 과정에 발생한 버그 수정

+++ src/main/java/net/abh0518/genetic/IGenetic.java
@@ -17,5 +17,5 @@ public interface IGenetic<T> {
 	 * @param mutationProb 돌연변이가 발생할 확률, 0.1이나 0.2 등으로 백분율 소수로 표현 한다.
 	 * @return
 	 */
-	List<T> geneticOptimize(int genSize, int maxLoop, double eliteRate, double mutationProb);
+	List<T> geneticOptimize(int genSize, int maxLoop, float eliteRate, float mutationProb);
 }
